Improve API validation and error handling

+++ src/api.py
@@ -1,8 +1,8 @@
-from fastapi import FastAPI
-from src import database
-import datetime
-from pydantic import BaseModel
+from dataclasses import field
 
+from fastapi import FastAPI,HTTPException
+from src import database, workouts
+from pydantic import BaseModel,Field
 app = FastAPI()
 
 
@@ -19,14 +19,16 @@ def get_workouts():
     ]
 @app.get("/workouts/{workout_id}")
 def get_workout(workout_id: int):
+    if workout_id not in workouts:
+        raise HTTPException(status_code=404, detail="Workout not found.")
     return [
         database.get_workout_by_id(workout_id)
     ]
 class Workout(BaseModel):
     cwiczenie: str
-    ciezar: int
-    powtorzenia: int
-    serie: int
+    ciezar: int = Field(gt=0)
+    powtorzenia: int = Field(gt=0)
+    serie: int = Field(gt=0)
 @app.post("/workouts")
 def post_workout(workout: Workout):
     workout={
@@ -34,17 +36,24 @@ def post_workout(workout: Workout):
             "ciezar": workout.ciezar,
             "powtorzenia": workout.powtorzenia,
             "serie": workout.serie,
-            "data": datetime.datetime.now().strftime("%Y-%m-%d %H:%M")
+            "data": workouts.data
         }
     database.add_workout(workout)
+    message = "Workout added successfully!"
+    return message
+
 
 @app.delete("/workouts/{workout_id}")
 def delete_workout(workout_id: int):
+    if workout_id not in workouts:
+        raise HTTPException(status_code=404, detail="Workout not found.")
     return [
         database.delete_workout(workout_id)
     ]
 @app.put("/workouts/{workout_id}")
 def put_workout(workout_id: int, column: str, value):
+    if workout_id not in workouts:
+        raise HTTPException(status_code=404, detail="Workout not found.")
     return [
         database.update_workout(
             workout_id,
